feat: change the functionality so that if a number is added to an already existing user, the new number will replace the old number

+++ part2/phonebook/src/components/Persons.jsx
@@ -6,6 +6,10 @@ function Persons({ persons, setPersons, filter }) {
   ) || [];
 
   const deletePerson = (id) => {
+    if (!window.confirm('Are you sure you want to delete this person?')) {
+      return;
+    }
+
     personService.deletePerson(id).then(() => {
       setPersons(persons.filter(person => person.id !== id));
       console.log(`Deleted person with id: ${id}`);
